pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/folio"]

[project]
name = "folio"
version = "0.1.0"
description = "Build beautiful self-contained HTML reports from Python analysis"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["data", "reporting", "html", "pandas", "pyspark", "visualization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "jinja2>=3.1",
    "pandas>=2.0",
    "markdown-it-py>=3.0",
]

[project.optional-dependencies]
# pip install folio[email]
email = [
    "premailer>=3.10",
    "kaleido>=0.2",
]
# pip install folio[dev]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-snapshot>=0.9",
    "beautifulsoup4>=4.12",
    "hypothesis>=6.100",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=folio --cov-report=term-missing -v"

[tool.coverage.run]
source = ["src/folio"]
omit = ["*/tests/*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
